requirements.txt
asgiref==3.2.3
Django==3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
Pillow==6.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
